Fixed a mistake.
I thought to acess profiles on beecrowd you needed to be logged in but apparently that is not actualy true. I had come to that conclusion because I tested if it was possible to acess my profile without being logged in and I wasn't able to but now I tested it again and I was able to.